Rename module as keep-alive is not its only function
