+++ repository-companion/src/main/resources/3RD-PARTY
