+++ repository-companion/src/main/resources/config.properties
