+++ repository-companion/src/main/resources/logback.xml
