+++ repository-companion/src/test/docker/alfresco/extension/dev-log4j.properties
