+++ repository-companion/src/test/docker/companion-01-custom-config.properties
