+++ repository-companion/src/test/docker/companion-02-custom-config.properties
