+++ repository-companion/src/test/docker/repository-1-logs/dummy.properties
