+++ repository-companion/src/test/docker/repository-2-logs/dummy.properties
